[refactor] 헥사고날  패키지 리팩터링
