+++ src/main/java/com/example/bolmal/auth/infrastructure/RefreshRepository.java
@@ -1,6 +1,6 @@
-package com.example.bolmal.config.repository;
+package com.example.bolmal.auth.infrastructure;
 
-import com.example.bolmal.config.domain.RefreshEntity;
+import com.example.bolmal.auth.domain.RefreshEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.transaction.annotation.Transactional;
