+++ src/main/java/com/example/bolmal/member/domain/Gender.java
@@ -1,4 +1,4 @@
-package com.example.bolmal.config.domain;
+package com.example.bolmal.member.domain;
 
 public enum Gender {
