+++ src/main/java/com/example/bolmal/member/domain/Status.java
@@ -1,4 +1,4 @@
-package com.example.bolmal.config.domain;
+package com.example.bolmal.member.domain;
 
 public enum Status {
     ACTIVE,INACTIVE
